% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_to_casrn.R
\name{convert_date_to_casrn}
\alias{convert_date_to_casrn}
\title{Convert Dates Back to CAS Registry Numbers}
\usage{
convert_date_to_casrn(date_str)
}
\arguments{
\item{date_str}{A character vector or a dataframe column containing combination of CASRNs and date strings.}
}
\value{
A character vector with dates converted back to CASRNs, where applicable.
}
\description{
This function converts date strings that were mistakenly converted from CAS registry numbers back to their original CAS registry number format.
}
\examples{
# Example with a vector
date_vector <- c("3/1/2038", "100-00-5", "3/7/2680", "10/9/3546", "79-06-1")
convert_date_to_casrn(date_vector)

# Example with a dataframe
df <- data.frame(
  id = 1:5,
  casrn = c("3/1/2038", "100-00-5", "3/7/2680", "10/9/3546", "79-06-1"),
  stringsAsFactors = FALSE
)
df$casrn <- convert_date_to_casrn(df$casrn)

}
